Add a new target for clients running at Android OS level
Introducing a new target that sets the ANDROID_OS flag.
The flag indicates that the library is built for a client
running at OS-level such as adbd.

This was added to provide a way for Android processes with
the necessary permisssions to listen to libusb hotplug events
through netlink.

Related discussions on the libusb repository:
- https://github.com/libusb/libusb/pull/1577
- https://github.com/libusb/libusb/pull/1168
- https://github.com/libusb/libusb/commit \
  /2f3bc98b0d0f4766496df53c855685a5f0e5e7cf

Test: TreeHugger
Change-Id: Ib93a0a246123eb2ecb630f79b318de2fdfa78ebb